pyproject.toml
[project]
name = "tbelt"
description = "format or run checks on files by running tools on them"
authors = [{ name = "hotdog-werx" }]
readme = "README.md"
license = 'MIT'
requires-python = ">=3.11"
dynamic = ["version"]

dependencies = [
    "pydantic>=2.11.7, <3",
    "pyyaml>=6.0, <7",
    "pathspec>=0.12.0, <1",
    "structlog>=25.4.0, <26",
    "rich>=14.0.0, <15",
]

[project.scripts]
tb = "toolbelt.cli.main:main"
tbelt = "toolbelt.cli.main:main"
toolbelt = "toolbelt.cli.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch]
build.packages = ["toolbelt"]
metadata.allow-direct-references = true

[tool.hatch.version]
path = 'toolbelt/version.py'

[tool.poe]
include = ["configs/tasks.toml"]

[tool.ruff]
extend = "configs/ruff.toml"
exclude = [
  "tests/**/fixtures/**",  # Exclude test fixture resources from linting
]

[tool.pyright]
typeCheckingMode = "standard"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.toolbelt]
include = ["@toolbelt:resources/presets/hdw.yaml", "toolbelt.yaml"]
